[#2639] Move ephemera in its own module
